+++ packages/blade/src/components/Box/Box.tsx
@@ -15,6 +15,7 @@ const validateBackgroundString = (stringBackgroundColorValue: string): void => {
     if (
       !stringBackgroundColorValue.startsWith('surface.background') &&
       !stringBackgroundColorValue.startsWith('overlay.') &&
+      !stringBackgroundColorValue.startsWith('feedback.background') &&
       stringBackgroundColorValue !== 'transparent'
     ) {
       throwBladeError({