+++ packages/blade/src/components/DatePicker/DatePicker.stories.tsx
@@ -63,6 +63,7 @@ export default {
     onPreviousMonth: baseProp,
     onPreviousYear: baseProp,
     locale: baseProp,
+    footer: baseProp,
     accessibilityLabel: inputProp,
     errorText: inputProp,
     helpText: inputProp,
@@ -111,11 +112,6 @@ import { m } from 'framer-motion';
 } as Meta<DatePickerProps<'single' | 'range'>>;
 
 const DatePickerTemplate: StoryFn<typeof DatePickerComponent> = ({ ...args }) => {
-  if (args.selectionType === 'range' && typeof args.label === 'string') {
-    throw new Error(
-      '[Storybook Controls]: Cannot use string label for range selection, please switch to the RangeDatePicker story',
-    );
-  }
   if (args.selectionType === 'single' && typeof args.label === 'object') {
     throw new Error(
       '[Storybook Controls]: Cannot use {start,end} label for single selection, please switch to the SingleDatePicker story',
@@ -137,16 +133,22 @@ SingleDatePicker.storyName = 'SingleDatePicker';
 SingleDatePicker.args = {
   label: 'Select a date',
   selectionType: 'single',
+  size: 'large',
 };
 
 export const RangeDatePicker = DatePickerTemplate.bind({});
 RangeDatePicker.storyName = 'RangeDatePicker';
 RangeDatePicker.args = {
-  label: { start: 'Start Date', end: 'End Date' },
+  label: { start: 'Select a date range' },
   selectionType: 'range',
 };
 
-export const DatePickerPresets: StoryFn<typeof DatePickerComponent> = ({ ...args }) => {
+export const DatePickerPresets: StoryFn<typeof DatePickerComponent> = ({ ..._args }) => {
+  const [selectedDates, setSelectedDates] = React.useState<DatesRangeValue>([
+    dayjs().subtract(7, 'days').toDate(),
+    dayjs().toDate(),
+  ]);
+
   return (
     <Box>
       <Text>
@@ -167,11 +169,22 @@ export const DatePickerPresets: StoryFn<typeof DatePickerComponent> = ({ ...args
       </Text>
 
       <DatePickerComponent
+        label={{ start: 'Select a date range' }}
         selectionType="range"
+        value={selectedDates}
         onChange={(date) => {
           console.log(date);
+          setSelectedDates(date as DatesRangeValue);
         }}
         presets={[
+          {
+            label: 'Today',
+            value: (date) => [dayjs(date).startOf('day').toDate(), date],
+          },
+          {
+            label: 'Yesterday',
+            value: (date) => [dayjs(date).subtract(1, 'day').startOf('day').toDate(), date],
+          },
           {
             label: 'Past 7 days',
             value: (date) => [dayjs(date).subtract(7, 'days').toDate(), date],
@@ -188,17 +201,26 @@ export const DatePickerPresets: StoryFn<typeof DatePickerComponent> = ({ ...args
             label: 'Past year',
             value: (date) => [dayjs(date).subtract(1, 'year').toDate(), date],
           },
+          {
+            label: 'Past financial year',
+            value: (date) => {
+              const d = dayjs(date);
+              const year = d.month() >= 3 ? d.year() : d.year() - 1;
+
+              return [dayjs(`${year - 1}-04-01`).toDate(), dayjs(`${year}-03-31`).toDate()];
+            },
+          },
+          {
+            label: 'Custom',
+            value: () => [null, null] as DatesRangeValue,
+          },
         ]}
-        {...args}
       />
     </Box>
   );
 };
 
 DatePickerPresets.storyName = 'With Presets';
-DatePickerPresets.args = {
-  label: { start: 'Start Date', end: 'End Date' },
-};
 
 export const DatePickerControlled: StoryFn<typeof DatePickerComponent> = () => {
   const [isOpen, setIsOpen] = React.useState(false);
@@ -222,7 +244,7 @@ export const DatePickerControlled: StoryFn<typeof DatePickerComponent> = () => {
         <Text marginTop="spacing.2">IsOpen: {JSON.stringify(isOpen)}</Text>
       </Box>
       <DatePickerComponent
-        label={{ start: 'Start Date', end: 'End Date' }}
+        label={{ start: 'Select a date range' }}
         selectionType="range"
         isOpen={isOpen}
         onOpenChange={({ isOpen }) => setIsOpen(isOpen)}
@@ -239,7 +261,7 @@ export const DatePickerControlled: StoryFn<typeof DatePickerComponent> = () => {
           selectionType="single"
           value={date}
           onChange={(date) => {
-            setDate(date);
+            if (date) setDate(date);
           }}
         />
       </Box>
@@ -278,8 +300,8 @@ export const Validations: StoryFn<typeof DatePickerComponent> = () => {
       </Text>
       <DatePickerComponent
         validationState={hasError ? 'error' : 'none'}
-        errorText={{ start: 'Cannot select a range which is more than 3 days' }}
-        label={{ start: 'Start Date', end: 'End Date' }}
+        errorText="Cannot select a range which is more than 3 days"
+        label={{ start: 'Select a date range' }}
         selectionType="range"
         value={date}
         onChange={(date) => {
@@ -310,7 +332,7 @@ export const MinMaxDates: StoryFn<typeof DatePickerComponent> = () => {
         <Text size="small">{`maxDate={dayjs().add(1, 'week').toDate()}`}</Text>
       </Box>
       <DatePickerComponent
-        label={{ start: 'Start Date', end: 'End Date' }}
+        label={{ start: 'Select a date range' }}
         selectionType="range"
         minDate={dayjs().subtract(1, 'week').toDate()}
         maxDate={dayjs().add(1, 'week').toDate()}
@@ -357,7 +379,7 @@ export const LabelPositionLeft: StoryFn<typeof DatePickerComponent> = () => {
         <DatePickerComponent
           labelPosition="left"
           selectionType="range"
-          label={{ start: 'Select a range' }}
+          label={{ start: 'Select a date range' }}
         />
         <DatePickerComponent selectionType="single" labelPosition="left" label="Select a date" />
       </Box>
@@ -385,7 +407,7 @@ export const MonthPicker: StoryFn<typeof DatePickerComponent> = ({ ...args }) =>
       >
         Note: picker is only supported in single selection mode
       </Text>
-      <DatePickerComponent picker="month" selectionType="single" {...args} />
+      <DatePickerComponent format="MMM" picker="month" selectionType="single" {...args} />
     </Box>
   );
 };
@@ -596,7 +618,7 @@ export const DatePickerWithLabelSuffixTrailing: StoryFn<typeof DatePickerCompone
         <DatePickerComponent
           labelPosition="left"
           selectionType="range"
-          label={{ start: 'Select a range' }}
+          label="Select a date range"
           labelSuffix={
             <Tooltip content="Select a date" placement="right">
               <TooltipInteractiveWrapper display="flex">
@@ -609,4 +631,30 @@ export const DatePickerWithLabelSuffixTrailing: StoryFn<typeof DatePickerCompone
       </Box>
     </Box>
   );
+};
+
+export const DatePickerWithFooter = DatePickerTemplate.bind({});
+DatePickerWithFooter.storyName = 'DatePicker with Footer';
+DatePickerWithFooter.args = {
+  label: 'Select a date',
+  selectionType: 'range',
+  footer: (
+    <Box display="flex" flexDirection="column" gap="spacing.2">
+      <Text size="small" color="surface.text.gray.normal">
+        This section only displays records from the last 45 days. This section only displays records
+        from the last 45 days.
+      </Text>
+      <Link size="small" href="#">
+        Link to report tab
+      </Link>
+    </Box>
+  ),
+};
+
+export const WithoutActionButtons = DatePickerTemplate.bind({});
+WithoutActionButtons.storyName = 'Without Action Buttons';
+WithoutActionButtons.args = {
+  label: 'Select a date',
+  selectionType: 'single',
+  showFooterActions: false,
 };