+++ packages/blade/src/components/Icons/CurrentAccountFilledIcon/CurrentAccountFilledIcon.native.test.tsx
@@ -1,0 +1,11 @@
+import CurrentAccountFilledIcon from '.';
+import renderWithTheme from '~utils/testing/renderWithTheme.native';
+
+describe('<CurrentAccountFilledIcon />', () => {
+  it('should render CurrentAccountFilledIcon', () => {
+    const renderTree = renderWithTheme(
+      <CurrentAccountFilledIcon color="feedback.icon.neutral.intense" size="large" />,
+    ).toJSON();
+    expect(renderTree).toMatchSnapshot();
+  });
+});