+++ packages/blade/src/components/Icons/LoansForBusinessesFilledIcon/LoansForBusinessesFilledIcon.web.test.tsx
@@ -1,0 +1,11 @@
+import LoansForBusinessesFilledIcon from './';
+import renderWithTheme from '~utils/testing/renderWithTheme.web';
+
+describe('<LoansForBusinessesFilledIcon />', () => {
+  it('should render LoansForBusinessesFilledIcon', () => {
+    const { container } = renderWithTheme(
+      <LoansForBusinessesFilledIcon color="feedback.icon.neutral.intense" size="large" />,
+    );
+    expect(container).toMatchSnapshot();
+  });
+});